database move to try catch

+++ src/app/(server)/cart/get-cart/[email]/route.js
@@ -3,9 +3,16 @@ import { NextResponse } from "next/server";
 
 //Get cart items
 export const GET = async (request, { params }) => {
-  const db = await connectDB();
-  const cartsCollection = await db.collection("carts");
   try {
+    const db = await connectDB();
+    const cartsCollection = await db.collection("carts");
+    //checking Email at params
+    if (!params?.email) {
+      return NextResponse.json(
+        { message: "Email parameter is missing" },
+        { status: 400 }
+      );
+    }
     const carts = await cartsCollection.find({ email: params.email }).toArray();
     return NextResponse.json(carts, { status: 200 });
   } catch (error) {
@@ -19,10 +26,17 @@ export const GET = async (request, { params }) => {
 
 //Delete cart items
 export const DELETE = async (request, { params }) => {
-  const db = await connectDB();
-  const cartsCollection = await db.collection("carts");
-
   try {
+    const db = await connectDB();
+    const cartsCollection = await db.collection("carts");
+
+    //checking Email at params
+    if (!params?.email) {
+      return NextResponse.json(
+        { message: "Email parameter is missing" },
+        { status: 400 }
+      );
+    }
     await cartsCollection.deleteMany({ email: params.email });
     return NextResponse.json({ message: "All items Deleted", status: 200 });
   } catch (error) {
